requirements.txt
streamlit>=1.28.0
pdfplumber>=0.7.0
pillow>=9.0.0
pyyaml>=6.0
python-docx>=0.8.11
openpyxl>=3.1.0
pandas>=1.5.0
openai>=1.0.0
python-dotenv>=1.0.0